# GH-37939: [C++] Use signed arithmetic for frame of reference when encoding DELTA_BINARY_PACKED