@mapleFU 2023-09-29
Oh after check arrow-rs( https://github.com/apache/arrow-rs/blob/471f6dd2911d8328ca56efe2f685e08c0a3fb8c8/parquet/src/encodings/encoding/mod.rs#L275 ) and parquet-mr ( See `DeltaBinaryPackingValuesWriterForLong.flushBlockBuffer` ), so we can just move forward, great