@pitrou 2023-10-03
@etseidl Did you try to read data generated by this patch using other Parquet implementations such as parquet-mr?