@etseidl 2023-10-03
> @etseidl Did you try to read data generated by this patch using other Parquet implementations such as parquet-mr?

The data I posted in the issue was readable by arrow-rs and parquet-mr, but I can do a bigger test of that with more varied data.