@pitrou 2023-10-03
> The data I posted in the issue was readable by arrow-rs and parquet-mr, but I can do a bigger test of that with more varied data.

No need to. Thanks for the answer!