@etseidl 2023-10-03
Thank you all for your help with this! Should I rebase and force push now? And should I update the description to match the current state of the patch (as it no longer promotes to a larger bit width)? @mapleFU @pitrou